from_server/requirements.txt
amqp==1.4.8
anyjson==0.3.3
billiard==3.3.0.22
celery==3.1.19
diff-match-patch==20121119
Django==1.9
django-celery==3.1.17
django-cors-headers==1.1.0
django-filter==0.11.0
django-import-export==0.4.2
django-rest-auth==0.6.0
djangorestframework==3.3.2
gunicorn==19.4.1
kombu==3.0.32
Markdown==2.6.5
MySQL-python==1.2.5
Pillow==3.0.0
Pygments==2.0.2
pytz==2015.7
six==1.10.0
tablib==0.10.0
wheel==0.24.0

